FOUR-24891 : FE: remove the fake variables related to the custom columns
