+++ resources/js/processes-catalogue/components/api/index.js
@@ -11,12 +11,8 @@ export const getRequests = async ({
       pmql,
       order_direction: orderDirection,
       order_by: orderBy,
-      // non_system: nonSystem,
-      // all_inbox: allInbox,
-      // processesIManage,
       filter,
       advanced_filter: JSON.stringify(advancedFilter),
-      // status_filter: statusFilter,
     },
   });
