+++ resources/js/processes-catalogue/components/home/config/index.js
@@ -1,4 +1,3 @@
 export * from "./columns";
 export * from "./filters";
-export * from "./requestDefaultColumns";
 export * from "./metrics";
