--- resources/js/processes-catalogue/components/home/config/requestDefaultColumns.js
@@ -1,162 +1,0 @@
-export const subjectColumns = [
-  {
-    label: "Request ID",
-    field: "id",
-    sortable: true,
-    default: true,
-    width: 115,
-    subject: {
-      type: "Field",
-      value: "id",
-    },
-  },
-  {
-    label: "Case #",
-    field: "case_number",
-    sortable: true,
-    default: true,
-    width: 95,
-    subject: {
-      type: "Field",
-      value: "case_number",
-    },
-  },
-  {
-    label: "Case title",
-    field: "case_title",
-    sortable: true,
-    default: true,
-    truncate: true,
-    width: 200,
-    subject: {
-      type: "Field",
-      value: "case_title",
-    },
-  },
-  {
-    label: "Process",
-    field: "name",
-    sortable: true,
-    default: true,
-    width: 200,
-    truncate: true,
-    subject: {
-      type: "Field",
-      value: "name",
-    },
-  },
-  {
-    label: "Stage",
-    field: "stage",
-    sortable: true,
-    default: true,
-    width: 200,
-    truncate: true,
-    subject: {
-      type: "Field",
-      value: "stage",
-    },
-  },
-  {
-    label: "Progress",
-    field: "progress",
-    sortable: true,
-    default: true,
-    width: 200,
-    truncate: true,
-    subject: {
-      type: "Field",
-      value: "progress",
-    },
-  },
-  {
-    label: "Participants",
-    field: "participants",
-    sortable: true,
-    default: true,
-    width: 200,
-    truncate: true,
-    subject: { type: "ParticipantsFullName" },
-    hideSortingButtons: true,
-  },
-  {
-    label: "Status",
-    field: "status",
-    sortable: true,
-    default: true,
-    width: 200,
-    subject: { type: "Status" },
-  },
-  {
-    label: "Started",
-    field: "initiated_at",
-    format: "datetime",
-    sortable: true,
-    default: true,
-    width: 200,
-    subject: {
-      type: "Field",
-      value: "initiated_at",
-    },
-  },
-  {
-    label: "Completed",
-    field: "completed_at",
-    format: "datetime",
-    sortable: true,
-    default: true,
-    width: 200,
-    subject: {
-      type: "Field",
-      value: "completed_at",
-    },
-  },
-  {
-    label: "Name",
-    field: "data.program.name",
-    sortable: true,
-    default: true,
-    width: 200,
-    truncate: true,
-    subject: {
-      type: "Field",
-      value: "data.program.name",
-    },
-  },
-  {
-    label: "Aid Type",
-    field: "data.program.type",
-    sortable: true,
-    default: true,
-    width: 200,
-    truncate: true,
-    subject: {
-      type: "Field",
-      value: "data.program.type",
-    },
-  },
-  {
-    label: "Source",
-    field: "data.program.source",
-    sortable: true,
-    default: true,
-    width: 200,
-    truncate: true,
-    subject: {
-      type: "Field",
-      value: "data.program.source",
-    },
-  },
-  {
-    label: "Deadline",
-    field: "data.program.deadline",
-    sortable: true,
-    default: true,
-    width: 200,
-    truncate: true,
-    subject: {
-      type: "Field",
-      value: "data.program.deadline",
-    },
-  },
-];
